[fix](group commit)Fix wal manager ut
