+++ be/test/http/stream_load_test.cpp
@@ -45,7 +45,10 @@ public:
     StreamLoadTest() = default;
     virtual ~StreamLoadTest() = default;
     void SetUp() override {}
-    void TearDown() override {}
+    void TearDown() override {
+        SAFE_STOP(ExecEnv::GetInstance()->_wal_manager);
+        ExecEnv::GetInstance()->clear_wal_mgr();
+    }
 };
 
 void http_request_done_cb(struct evhttp_request* req, void* arg) {
@@ -58,7 +61,7 @@ TEST_F(StreamLoadTest, TestHeader) {
     static_cast<void>(wal_mgr->_wal_dirs_info->add("test_path_1", 1000, 0, 0));
     static_cast<void>(wal_mgr->_wal_dirs_info->add("test_path_2", 10000, 0, 0));
     static_cast<void>(wal_mgr->_wal_dirs_info->add("test_path_3", 100000, 0, 0));
-    ExecEnv::GetInstance()->set_wal_mgr(wal_mgr);
+    ExecEnv::GetInstance()->set_wal_mgr(std::move(wal_mgr));
     // 1. empty info
     {
         auto* evhttp_req = evhttp_request_new(nullptr, nullptr);
